后端：把图片表拆分为图片表和子图片表并重构对应逻辑，前端：优化用户体验

+++ src/FishPics-backend/src/main/java/hk/ljx/fishpicsbackend/dto/post/EditPostRequest.java
@@ -33,9 +33,9 @@ public class EditPostRequest implements Serializable {
     private String content;
 
     /**
-     * 封面图片的 id
+     * 封面图片的 id（图片列表索引）
      */
-    private Long cover;
+    private Integer cover;
 
     /**
      * 0-公开，1-仅自己可见，
